Continuazione BE. Rilascio ricerca libri

+++ backend/LibraryService/src/main/java/it/socialibrary/libraryservice/service/impl/BookService.java
@@ -10,11 +10,12 @@ import it.socialibrary.libraryservice.web.dto.BookDto;
 import it.socialibrary.libraryservice.web.dto.FiltersBookDto;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
-import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.domain.Specification;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
+import java.io.IOException;
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
@@ -23,17 +24,45 @@ import java.util.UUID;
 public class BookService implements IBookService {
 
     private final BookRepository bookRepository;
+
+    private final BookImageService bookImageService;
     private final IBookMapper bookMapper;
 
     @Autowired
-    public BookService(BookRepository bookRepository, CategoryService categoryService, IBookMapper bookMapper) {
+    public BookService(BookRepository bookRepository, BookImageService bookImageService, IBookMapper bookMapper) {
         this.bookRepository = bookRepository;
+        this.bookImageService = bookImageService;
         this.bookMapper = bookMapper;
     }
 
+    /**
+     * Ricerca di libri in base ai filtri forniti.
+     * @param filters Filtri
+     * @param pageable Paginazione
+     * @return Lista di libri filtrati
+     */
     @Override
     public Page<BookDto> searchBookOnFilter(FiltersBookDto filters, Pageable pageable) {
         Specification<Book> spec = BookSpecification.searchBook(filters);
+
+        if(filters.getLongitude() != null && filters.getLatitude() != null && filters.getRadius() != null) {
+            try {
+                int radiusInMeters = filters.getRadius() * 1000;
+
+                // Ricerco gli ID dei libri che si trovano all'interno di un certo raggio in base alle coordinate fornite
+                List<UUID> findIdsForDistance = this.bookRepository.findDistance(
+                        filters.getLongitude(), filters.getLatitude(), radiusInMeters);
+
+                if(findIdsForDistance.isEmpty()) return Page.empty(pageable);
+
+                // Aggiungo la condizione di ricerca dei libri trovati all'interno del raggio'
+                spec = spec.and(((root, cq, cb) ->
+                        root.get("id").in(findIdsForDistance))
+                );
+            } catch (Exception e) {
+                throw new IllegalArgumentException("Invalid coordinates or radius");
+            }
+        }
 
         return bookRepository.findAll(spec, pageable).map(bookMapper::toDTO);
     }
@@ -41,24 +70,43 @@ public class BookService implements IBookService {
     @Override
     public BookDto findById(UUID id) throws NotFoundException {
         Optional<Book> book = this.bookRepository.findById(id);
-        // Se il libro non esiste, lancio un'eccezione'
-        book.orElseThrow(() -> new NotFoundException("Record not found"));
+        if(book.isEmpty()) throw new NotFoundException("Record not found");
 
         return this.bookMapper.toDTO(book.get());
     }
 
     @Override
-    public void save(BookDto bookDto) {
-        this.bookRepository.save(this.bookMapper.toEntity(bookDto));
+    public Page<BookDto> findByUser(UUID userId, Pageable pageable) {
+        Page<Book> book = this.bookRepository.findByUserId(userId, pageable);
+        return book.map(this.bookMapper::toDTO);
+    }
+
+
+    @Override
+    @Transactional
+    public void save(BookDto bookDto) throws IOException {
+        Book book = this.bookMapper.toEntity(bookDto);
+        this.bookRepository.save(book);
+
+        // TODO: GESTIONE DELLE IMMAGINI
+        this.bookImageService.createThumbnail(book, null);
+        this.bookImageService.createPreview(book, null);
     }
 
     @Override
+    @Transactional
     public BookDto update(BookDto bookDto) {
-        return null;
+        Book book = this.bookMapper.toEntity(bookDto);
+        this.bookRepository.save(book);
+
+        return findById(bookDto.getId());
     }
 
     @Override
-    public void delete(UUID id) {
+    @Transactional
+    public void delete(UUID id) throws NotFoundException {
+        if(!this.bookRepository.existsById(id)) throw new NotFoundException("Record not found");
+
         this.bookRepository.deleteById(id);
     }
 }
